Refactor to just use many seperate golang packages.
