+++ .gitignore
@@ -5,3 +5,4 @@ cri-resource-manager-0.6.0.x86_64.tar.gz
 /test/e2e/cri-resmgr-extension/cri-resmgr-extension.test
 /gardener-extension-cri-resmgr.e2e-tests
 /gardener-extension-cri-resmgr.test
+gardener-extension-cri-resmgr.actuator.test
